Split v1.py into smaller, more-specific route files
